+++ app/Http/Controllers/Chart/TagReportController.php
@@ -450,8 +450,6 @@ class TagReportController extends Controller
         $data = $this->generator->multiCurrencyPieChart($result);
 
         return response()->json($data);
-
-        return response()->json($data);
     }