+++ resources/lang/de_DE/firefly.php
@@ -1451,6 +1451,7 @@ return [
     'tools_index_intro'                     => 'Es gibt mehrere Werkzeuge, um Daten in Firefly III zu importieren (Diese werden unten vorgestellt). Weitere Informationen finden Sie unter <a href="https://docs.firefly-iii.org/importing-data/introduction">hier auf dieser Seite</a>.',
     'firefly_iii_csv_importer_name'         => 'Firefly III CSV-Import',
     'firefly_iii_bunq_importer_name'        => 'Firefly III Bunq 🌈 importieren',
+    'firefly_iii_ynab_importer_name'        => 'Firefly III YNAB importer',
     'ludo_revolut_importer_name'            => 'Ludo444\'s Revolut-Importer',
     //
     // sandstorm.io errors and messages:
@@ -1604,7 +1605,7 @@ return [
     'telemetry_admin_overview'           => 'Telemetrieübersicht',
     'telemetry_back_to_index'            => 'Zurück zum Telemetrieindex',
     'not_yet_submitted'                  => 'Noch nicht übermittelt',
-    'telemetry_type_feature'             => 'Feature flag',
+    'telemetry_type_feature'             => 'Funktions-Flag',
     'telemetry_submit_all'               => 'Datensätze übermitteln',
     'telemetry_delete_submitted_records' => 'Übertragene Datensätze löschen',
     'telemetry_submission_executed'      => 'Datensätze wurden übermittelt. Überprüfen Sie Ihre Protokolldateien für weitere Informationen.',